requirements.txt
flask
requests
gspread
oauth2client
coinbase-advanced-trade